Added end to end process querying logic and successful optimization

+++ src/Workloads/SimpleWorkload_1/Program.cs
@@ -1,5 +1,6 @@
 ﻿using CommandLine;
 
+// A "smoke-test" of a workload that sleeps for the least amount of time if the input to the program is in the range [1., 2).
 namespace SimpleWorkload_1
 {
     class Program
@@ -10,7 +11,7 @@ namespace SimpleWorkload_1
         public class Options
         {
             [Option('i', "input", Required = true, HelpText = "An int that will correspond to a sleep time.")]
-            public int Input { get; set; }
+            public double Input { get; set; }
         }
 
         static void Main(string[] args)
@@ -19,7 +20,7 @@ namespace SimpleWorkload_1
             {
                 switch (o.Input)
                 {
-                    case int n when n >= 1.0 && n < 2:
+                    case double n when n >= 1.0 && n < 2.0:
                         {
                             Thread.Sleep(FAST_SLEEP_MSEC);
                             break;
